+++ test/Command/InitCommandTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace ShlinkioTest\Shlink\Installer\Command;
 
+use PHPUnit\Framework\Assert;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\MockObject\MockObject;
@@ -15,7 +16,6 @@ use Symfony\Component\Console\Application;
 use Symfony\Component\Console\Tester\CommandTester;
 
 use function count;
-use function Functional\contains;
 
 class InitCommandTest extends TestCase
 {
@@ -34,14 +34,19 @@ class InitCommandTest extends TestCase
     }
 
     #[Test, DataProvider('provideInputs')]
-    public function expectedCommandsAreRunBasedOnInput(array $input, array $expectedCommands): void
+    public function expectedCommandsAreRunBasedOnInput(array $input, array $commands, bool $interactive): void
     {
-        $this->commandsRunner->expects($this->exactly(count($expectedCommands)))->method('execPhpCommand')->with(
-            $this->callback(fn (string $commandName) => contains($expectedCommands, $commandName)),
-            $this->anything(),
-        )->willReturn(true);
+        $this->commandsRunner->expects($this->exactly(count($commands)))->method('execPhpCommand')->willReturnCallback(
+            function (string $commandName, $_, bool $isInteractive, array $args) use ($commands, $interactive): bool {
+                Assert::assertContains($commandName, $commands);
+                Assert::assertEquals($interactive, $isInteractive);
+                Assert::assertEquals($commandName === InstallationCommand::API_KEY_CREATE->value ? ['foo'] : [], $args);
 
-        $this->tester->execute($input);
+                return true;
+            },
+        );
+
+        $this->tester->execute($input, ['interactive' => $interactive]);
     }
 
     public static function provideInputs(): iterable
@@ -51,14 +56,20 @@ class InitCommandTest extends TestCase
             InstallationCommand::DB_MIGRATE->value,
             InstallationCommand::ORM_PROXIES->value,
             InstallationCommand::GEOLITE_DOWNLOAD_DB->value,
-        ]];
+        ], true];
+        yield 'non-interactive' => [[], [
+            InstallationCommand::DB_CREATE_SCHEMA->value,
+            InstallationCommand::DB_MIGRATE->value,
+            InstallationCommand::ORM_PROXIES->value,
+            InstallationCommand::GEOLITE_DOWNLOAD_DB->value,
+        ], false];
         yield 'skips' => [['--skip-initialize-db' => true, '--skip-download-geolite' => true], [
             InstallationCommand::DB_MIGRATE->value,
             InstallationCommand::ORM_PROXIES->value,
-        ]];
+        ], true];
         yield 'all' => [[
             '--clear-db-cache' => true,
-            '--initial-api-key' => true,
+            '--initial-api-key' => null,
             '--download-rr-binary' => true,
         ], [
             InstallationCommand::DB_CREATE_SCHEMA->value,
@@ -68,16 +79,25 @@ class InitCommandTest extends TestCase
             InstallationCommand::GEOLITE_DOWNLOAD_DB->value,
             InstallationCommand::API_KEY_GENERATE->value,
             InstallationCommand::ROAD_RUNNER_BINARY_DOWNLOAD->value,
-        ]];
+        ], true];
         yield 'mixed' => [[
-            '--initial-api-key' => true,
+            '--initial-api-key' => null,
             '--skip-download-geolite' => true,
         ], [
             InstallationCommand::DB_CREATE_SCHEMA->value,
             InstallationCommand::DB_MIGRATE->value,
             InstallationCommand::ORM_PROXIES->value,
             InstallationCommand::API_KEY_GENERATE->value,
-        ]];
+        ], true];
+        yield 'api key value' => [[
+            '--initial-api-key' => 'foo',
+        ], [
+            InstallationCommand::DB_CREATE_SCHEMA->value,
+            InstallationCommand::DB_MIGRATE->value,
+            InstallationCommand::ORM_PROXIES->value,
+            InstallationCommand::API_KEY_CREATE->value,
+            InstallationCommand::GEOLITE_DOWNLOAD_DB->value,
+        ], true];
     }
 
     #[Test, DataProvider('provideExitCodes')]